+++ plone/app/content/item.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from Products.CMFCore.PortalContent import PortalContent
-from Products.CMFDefault.DublinCore import DefaultDublinCoreImpl
+from Products.CMFPlone.DublinCore import DefaultDublinCoreImpl
 from plone.app.content.interfaces import IReindexOnModify
 from zope.container.contained import Contained
 from zope.interface import implementer